Refactor diff logic into separate file with exported API
